# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Poachy API Documentation",
        "description": "Multi-tenant POS and eCommerce Marketplace Platform API",
        "contact": {
            "name": "Poachy Support",
            "email": "support@poachy.com"
        },
        "license": {
            "name": "Proprietary",
            "url": "https://poachy.com/license"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/v1/central/business-details/pending": {
            "get": {
                "tags": [
                    "Business Review"
                ],
                "summary": "Get pending business details",
                "description": "Admin retrieves all business details awaiting approval",
                "operationId": "5abfd6c2121bbf5f00785f0699763711",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page",
                        "schema": {
                            "type": "integer",
                            "default": 15,
                            "maximum": 100
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pending business details retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Pending business details retrieved successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden - Admin only"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/central/business-details": {
            "get": {
                "tags": [
                    "Business Review"
                ],
                "summary": "Get all business details",
                "description": "Admin retrieves all business details with optional filters",
                "operationId": "7424ea45fdc56c069f4e7d77f558d915",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter by status",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "pending",
                                "active",
                                "inactive",
                                "suspended"
                            ]
                        }
                    },
                    {
                        "name": "business_type_id",
                        "in": "query",
                        "description": "Filter by business type",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "is_verified",
                        "in": "query",
                        "description": "Filter by verification status",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "city",
                        "in": "query",
                        "description": "Filter by city",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page",
                        "schema": {
                            "type": "integer",
                            "default": 15,
                            "maximum": 100
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Business details retrieved successfully"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden - Admin only"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/central/business-details/{id}/approve": {
            "post": {
                "tags": [
                    "Business Review"
                ],
                "summary": "Approve business details",
                "description": "Admin approves a business details submission",
                "operationId": "4594bc797545f7ba9f11ef3db7ebb55b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Business detail ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "notes": {
                                        "type": "string",
                                        "example": "Approved after review"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Business details approved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Business details approved successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/BusinessDetailResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden - Admin only"
                    },
                    "404": {
                        "description": "Business details not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/central/business-details/{id}/reject": {
            "post": {
                "tags": [
                    "Business Review"
                ],
                "summary": "Reject business details",
                "description": "Admin rejects a business details submission",
                "operationId": "1cb3be35a5d2a00ebda7a2ddc6119cad",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Business detail ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "notes": {
                                        "type": "string",
                                        "example": "Missing required documents"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Business details rejected successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Business details rejected successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden - Admin only"
                    },
                    "404": {
                        "description": "Business details not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "BusinessDetailResource": {
                "title": "Business Detail Resource",
                "description": "Business detail resource representation",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "tenant_id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "business_name": {
                        "type": "string",
                        "example": "Tech Haven Electronics"
                    },
                    "business_description": {
                        "type": "string",
                        "nullable": true
                    },
                    "business_logo": {
                        "type": "string",
                        "nullable": true
                    },
                    "business_banner": {
                        "type": "string",
                        "nullable": true
                    },
                    "business_type": {
                        "type": "object"
                    },
                    "business_category": {
                        "type": "object"
                    },
                    "business_email": {
                        "type": "string",
                        "nullable": true
                    },
                    "business_phone": {
                        "type": "string"
                    },
                    "contact_person": {
                        "type": "string",
                        "nullable": true
                    },
                    "address": {
                        "type": "string",
                        "nullable": true
                    },
                    "city": {
                        "type": "string",
                        "nullable": true
                    },
                    "county": {
                        "type": "string",
                        "nullable": true
                    },
                    "operating_hours": {
                        "type": "object",
                        "nullable": true
                    },
                    "delivery_info": {
                        "type": "object",
                        "nullable": true
                    },
                    "settings": {
                        "type": "object",
                        "nullable": true
                    },
                    "social_media": {
                        "type": "object",
                        "nullable": true
                    },
                    "rating": {
                        "type": "number",
                        "format": "float"
                    },
                    "rating_count": {
                        "type": "integer"
                    },
                    "is_verified": {
                        "type": "boolean"
                    },
                    "verified_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "status": {
                        "type": "string",
                        "example": "pending"
                    },
                    "onboarded_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Laravel Sanctum token authentication. Enter your token in the format: Bearer {token}",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}